pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recipes-bot"
version = "0.1.0"
description = "TikTok video downloader and recipes bot"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=2.2.0",
    "openai-whisper>=20231117",
    "playwright>=1.48.0",
    "python-telegram-bot>=21.6",
    "requests>=2.32.4",
    "yt-dlp>=2023.12.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[project.scripts]
tiktok-downloader = "recipes_bot.tiktok_downloader:main"
recipes-bot-run = "recipes_bot.bot:main"

[tool.setuptools]
packages = ["recipes_bot"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]
